Crate device_manager

Structs

FatFsAdapter

An adapter (wrapper type) that implements traits required by the [fatfs] crate for any I/O device that wants to be usable by [fatfs].

FatFsIoErrorAdapter

This struct exists to enable us to implement the [fatfs::IoError] trait for the [core2::io::Error] trait.

Functions

early_init

Performs early-stage initialization for simple devices needed during early boot.

init

Initializes all other devices not initialized during early_init().