Function device_manager::early_init

pub fn early_init(
    kernel_mmi: &mut MemoryManagementInfo
) -> Result<(), &'static str>

Performs early-stage initialization for simple devices needed during early boot.

This includes:

local APICs (apic),

acpi tables for system configuration info, including the IOAPIC.