Struct device_manager::FatFsAdapter

pub struct FatFsAdapter<IO>(_);

An adapter (wrapper type) that implements traits required by the [fatfs] crate for any I/O device that wants to be usable by [fatfs].

To meet [fatfs]’s requirements, the underlying I/O stream must be able to read, write, and seek while tracking its current offset. We use traits from the [core2] crate to meet these requirements, thus, the given IO parameter must implement those [core2] traits.

For example, this allows one to access a FAT filesystem by reading from or writing to a storage device.

Implementations

impl<IO> FatFsAdapter<IO>

pub fn new(io: IO) -> FatFsAdapter<IO>